create user POST Request implemented
running the post request from the route createuser on the Postman application makes a post request to create the new user (with the default static values found in the controllers\queryController.ts file. This post requst creates the new user using a mysql query in the mysql database. Do Note: the onclicklistener to perform this function is NOT working yet, that will be implemented next.